requirements.txt
streamlit==1.24.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
pywavelets==1.4.1
matplotlib==3.7.1
